added biblatex support and move main.tex|.bib|.pdf to main folder

+++ routes/draftJS.js
@@ -29,7 +29,7 @@ function sendFile(request, response, fileExtension, contentType) {
 			const username = user.username
 
 			console.log(`sending ${fileExtension}...\n`)
-			const tempFile = `./latex/${username}/main.${fileExtension}`;
+			const tempFile = `./latex/${username}/main/main.${fileExtension}`;
 			fs.readFile(tempFile, function (err, data) {
 				response.contentType(contentType);
 				response.send(data)
@@ -55,15 +55,93 @@ router.get('/tex',auth.verifyUser, (req, res, next) => {
 	sendFile(req, res, 'tex', 'application/x-tex')
 })
 
+router.post('/tex', auth.verifyUser, (req, res) => {
+	let body = req.body
+	console.log(body)
+
+	findUser(req)
+		.then(function (user) {
+			// Do something with the user
+			const username = user.username
+			fs.mkdirSync(`./latex/${username}`, { recursive: true })
+
+			fs.writeFile(`./latex/${username}/main.bib`, '', err => {
+				if (err) {
+					console.log('new file err', err)
+					return
+				}
+				console.log('writing...')
+			})
+
+			for (let item of Object.values(body)) {
+				console.log('hhhhhhhhh, Sync.........')
+				try {
+					fs.appendFileSync(
+						`./latex/${username}/main.bib`,
+						item + '\r\n' + '\r\n',
+					)
+				} catch (err) {
+					console.log('append err', err)
+					return
+				}
+			}
+
+			res.json({
+				status: 'success',
+				body: body
+			})
+		});
+})
+
 router.post('/', auth.verifyUser, (req, res, next) => {
 	findUser(req)
 		.then(function (user) {
 		// Do something with the user
 		const username = user.username
+		let isBib = false
 		fs.mkdirSync(`./latex/${username}`, { recursive: true })
 
-		empty(username)
-		setTimeout(() => insertBegin(username), 3000)
+		try {
+			if (fs.existsSync(`./latex/${username}/main.bib`)) {
+				isBib = true
+				console.log('-----------------main.bib--------------')
+			} else {
+				isBib = false
+				console.log('--------------no bib here--------------')
+			}
+		} catch(err) {
+			console.error(err)
+		}
+		console.log(isBib)
+		const begin = isBib ?
+			'\\documentclass[12pt]{article}' +
+			'\r\n' +
+			'\\usepackage[UTF8, scheme = plain, heading = false]{ctex}' +
+			'\r\n' +
+			'\\usepackage[left=20mm,right=20mm,top=25mm, bottom=15mm]{geometry}' +
+			'\r\n' +
+			'\\usepackage[colorlinks=true,pdfstartview=FitH,linkcolor=blue,anchorcolor=violet,citecolor=magenta]{hyperref}' +
+			'\r\n' +
+			'\\usepackage[backend=biber,style=gb7714-2015]{biblatex}' +
+			'\r\n' +
+			'\\addbibresource[location=local]{./main.bib}' +
+			'\r\n' +
+			'\\begin{document}' +
+			'\r\n'
+			:
+			'\\documentclass[12pt]{article}' +
+			'\r\n' +
+			'\\usepackage[UTF8, scheme = plain, heading = false]{ctex}' +
+			'\r\n' +
+			'\\begin{document}' +
+			'\r\n'
+		const end = isBib
+			? '\\printbibliography[heading=bibliography,title=参考文献]\r\n\\end{document}'
+			: '\\end{document}'
+
+		empty(username, isBib)
+
+		setTimeout(() => insertBegin(username, isBib, begin, end), 3000)
 		// return res.send(200);
 		});
 	/**
@@ -74,17 +152,9 @@ router.post('/', auth.verifyUser, (req, res, next) => {
 	 * await -> empty `latex` dir
 	 */
 	let body = req.body
-	const begin =
-		'\\documentclass[12pt]{article}' +
-		'\r\n' +
-		'\\usepackage[UTF8, scheme = plain, heading = false]{ctex}' +
-		'\r\n' +
-		'\\begin{document}' +
-		'\r\n'
-	const end = '\\end{document}'
 	console.log(body)
 
-	function empty(user) {
+	function empty(user, isBib) {
 		console.log('[1/5] empty ...........')
 		const dir = './latex/' + user
 
@@ -95,14 +165,27 @@ router.post('/', auth.verifyUser, (req, res, next) => {
 				if (err) return
 
 				for (const file of files) {
-					fs.unlink(path.join(dir, file), err1 => {
-						if (err1) throw err1
-					})
+					const filename = path.basename(file)
+					if (isBib === true) {
+						if (filename !== 'main.tex' && filename !== 'main.bib' && filename !== 'main') {
+							fs.unlink(path.join(dir, file), err1 => {
+								console.log(filename, 'has been removed!')
+								if (err1) throw err1
+							})
+						}
+					} else {
+						if (filename !== 'main.tex' && filename !== 'main') {
+							fs.unlink(path.join(dir, file), err1 => {
+								console.log(filename, 'has been removed!')
+								if (err1) throw err1
+							})
+						}
+					}
 				}
 			})
 	}
 
-	function insertBegin(user) {
+	function insertBegin(user, isBib, begin, end) {
 		console.log('[2/5] insertBegin........')
 		console.log('-----------begin-----------\n')
 		fs.writeFile(
@@ -113,12 +196,12 @@ router.post('/', auth.verifyUser, (req, res, next) => {
 					console.log('begin err', err)
 					return
 				}
-				appendContent(user)
+				appendContent(user, isBib, end)
 			}
 		)
 	}
 
-	function appendContent(user) {
+	function appendContent(user, isBib, end) {
 		console.log('[3/5] appendContent ............')
 		for (let i = 0; i < body.length; i += 1) {
 			try {
@@ -131,10 +214,10 @@ router.post('/', auth.verifyUser, (req, res, next) => {
 				return
 			}
 		}
-		appendEnd(user)
+		appendEnd(user, isBib, end)
 	}
 
-	function appendEnd(user) {
+	function appendEnd(user, isBib, end) {
 		console.log('[4/5] appendEnd ................\n')
 		try {
 			fs.appendFileSync(`./latex/${user}/main.tex`, end, )
@@ -145,12 +228,42 @@ router.post('/', auth.verifyUser, (req, res, next) => {
 
 		console.log('\n----------- end -----------\n')
 
-		compileTeX(user)
+		compileTeX(user, isBib)
 	}
 
-	function compileTeX(user) {
+	function compileTeX(user, isBib) {
 		console.log('[5/5] compile ..................')
 		console.log('...xelatex is running...\n')
+
+		function RenameFiles(files) {
+			for (let i = 0; i < files.length; i += 1) {
+				fs.rename(`./latex/${user}/${files[i]}`, `./latex/${user}/main/${files[i]}`, (err) => {
+					if (err) throw err;
+					console.log(`Rename ${files[i]} complete!`);
+				});
+			}
+		}
+
+		isBib ?
+			exec(`cd ./latex/${user} && latexmk -xelatex main.tex`, ((error, stdout, stderr) => {
+				if (error instanceof Error) {
+					console.log('latexmk error: ', error)
+					throw error
+				}
+				console.log('stdout: \n', stdout)
+				console.log('stderr: \n', stderr)
+				console.log('...latexmk finished...\n')
+				fs.mkdirSync(`./latex/${user}/main`, { recursive: true })
+				let moveFiles = ['main.tex', 'main.bib', 'main.pdf']
+
+				RenameFiles(moveFiles)
+
+				res.json({
+					status: 'success',
+					body: body
+				})
+			}))
+			:
 		exec(`cd ./latex/${user} && xelatex main.tex`, ((error, stdout, stderr) => {
 			if (error instanceof Error) {
 				console.log('latex error: ', error)
@@ -159,6 +272,11 @@ router.post('/', auth.verifyUser, (req, res, next) => {
 			console.log('stdout: \n', stdout)
 			console.log('stderr: \n', stderr)
 			console.log('...xelatex finished...\n')
+			fs.mkdirSync(`./latex/${user}/main`, { recursive: true })
+
+			let moveFiles = ['main.tex', 'main.pdf']
+
+			RenameFiles(moveFiles)
 
 			res.json({
 				status: 'success',
